+++ infra/uhrp-server-basic/src/routes/renew.ts
@@ -29,16 +29,9 @@ interface RenewResponse {
 }
 
 const renewHandler = async (req: RenewRequest, res: Response<RenewResponse>) => {
-  if (!req.auth.identityKey || req.auth.identityKey === 'unknown') {
-      return res.status(400).json({
-        status: 'error',
-        code: 'ERR_MISSING_IDENTITY_KEY',
-        description: 'Missing authfetch identityKey.'
-      })
-    }
   try {
     const { identityKey } = req.auth
-    if (!identityKey) {
+    if (!identityKey || identityKey === 'unknown') {
       return res.status(400).json({
         status: 'error',
         code: 'ERR_MISSING_IDENTITY_KEY',
@@ -70,7 +63,7 @@ const renewHandler = async (req: RenewRequest, res: Response<RenewResponse>) => 
     // Convert to MS to create an ISO string
     const newExpiryTimeSeconds = prevExpiryTime + (additionalMinutes * 60)
 
-    const fileSizeNum = parseInt(size, 10) || 0
+    const fileSizeNum = Number.parseInt(size, 10) || 0
     let amount = 0
     if (fileSizeNum > 0) {
       amount = await getPriceForFile({
@@ -108,7 +101,7 @@ const renewHandler = async (req: RenewRequest, res: Response<RenewResponse>) => 
       const expiryTag = out.tags.find(t => t.startsWith('expiry_time_'))
       if (!expiryTag) continue
 
-      const expiryNum = parseInt(expiryTag.substring('expiry_time_'.length), 10) || 0
+      const expiryNum = Number.parseInt(expiryTag.substring('expiry_time_'.length), 10) || 0
 
       if (expiryNum > maxpiry) {
         maxpiry = expiryNum
@@ -148,14 +141,16 @@ const renewHandler = async (req: RenewRequest, res: Response<RenewResponse>) => 
     )
 
     // Creating new tags
-    const newTags = []
+    const newTags: string[] = []
     if (prevAdvertisement.tags) {
       const uploaderTag = prevAdvertisement.tags.find(t => t.startsWith('uploader_identity_key_'))
       if (uploaderTag) newTags.push(uploaderTag)
     }
-    newTags.push(`uhrp_url_${Utils.toHex(Utils.toArray(uhrpUrl, 'utf8'))}`)
-    newTags.push(`object_identifier_${Utils.toHex(Utils.toArray(objectIdentifier, 'utf8'))}`)
-    newTags.push(`expiry_time_${newExpiryTimeSeconds}`)
+    newTags.push(
+      `uhrp_url_${Utils.toHex(Utils.toArray(uhrpUrl, 'utf8'))}`,
+      `object_identifier_${Utils.toHex(Utils.toArray(objectIdentifier, 'utf8'))}`,
+      `expiry_time_${newExpiryTimeSeconds}`
+    )
 
     const { signableTransaction } = await wallet.createAction({
       inputBEEF: BEEF,