+++ infra/uhrp-server-cloud-bucket/src/utils/getMetadata.ts
@@ -43,7 +43,7 @@ export async function getMetadata(uhrpUrl: string, uploaderIdentityKey: string, 
     const expiryTag = out.tags.find(t => t.startsWith('expiry_time_'))
     if (!objectIdTag || !expiryTag) continue
 
-    const expiryNum = parseInt(expiryTag.substring('expiry_time_'.length), 10) || 0
+    const expiryNum = Number.parseInt(expiryTag.substring('expiry_time_'.length), 10) || 0
 
     if (expiryNum > maxpiry) {
       maxpiry = expiryNum